requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyMuPDF
python-multipart
openai
streamlit
pdfplumber
pydantic
